Fix cardano-testnet

+++ cardano-testnet/src/Testnet/Defaults.hs
@@ -422,7 +422,7 @@ eraToProtocolVersion =
 -- TODO: Expose from cardano-api
 mkProtVer :: (Natural, Natural) -> ProtVer
 mkProtVer (majorProtVer, minorProtVer) =
-  case (`ProtVer` minorProtVer) <$> Ledger.mkVersion majorProtVer of
+  case (`ProtVer` fromIntegral minorProtVer) <$> Ledger.mkVersion majorProtVer of
     Just pVer -> pVer
     Nothing -> error "mkProtVer: invalid protocol version"
